refactor: reorganize component folder structure
