+++ src/components/HomeScreen/ComicGallery/ComicGallery.css
