+++ src/components/HomeScreen/ComicGallery/ComicGallery.tsx
@@ -1,7 +1,7 @@
 import { useNavigate } from "react-router-dom";
 import { useState, useEffect } from "react";
-import { apiBaseUrl } from "../../constants";
-import comicService from "../../services/comicService";
+import { apiBaseUrl } from "../../../constants";
+import comicService from "../../../services/comicService";
 import "./ComicGallery.css";
 import axios from "axios";
